add & use demo sub-module (move downloads)
