+++ specparam/demo/__init__.py
@@ -1,1 +1,3 @@
 """Demo sub-module."""
+
+from .download import load_example_data
